// back-end/prisma/generated/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  ADMIN: 'ADMIN',
  USER: 'USER'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const LotType = {
  SESSION_LINE: 'SESSION_LINE',
  TRANSACTION: 'TRANSACTION',
  LOT: 'LOT'
} as const

export type LotType = (typeof LotType)[keyof typeof LotType]


export const TransactionType = {
  PURCHASE: 'PURCHASE',
  FOUND: 'FOUND',
  GIFT: 'GIFT',
  EXISTING_STOCK: 'EXISTING_STOCK',
  LOST: 'LOST',
  GIVEN: 'GIVEN'
} as const

export type TransactionType = (typeof TransactionType)[keyof typeof TransactionType]


export const TransactionStatus = {
  SOLDED: 'SOLDED',
  RETURNED: 'RETURNED',
  RUNNING: 'RUNNING'
} as const

export type TransactionStatus = (typeof TransactionStatus)[keyof typeof TransactionStatus]


export const TransactionLineType = {
  GAIN: 'GAIN',
  LOST: 'LOST'
} as const

export type TransactionLineType = (typeof TransactionLineType)[keyof typeof TransactionLineType]